Integrated with Database file
The Checkout Basket can now calculate totals by applying a single
discount from the Discounts database.

It also has implemented BasketItem and PricedItem.

A BasketItem represents an item and it's quantity.
A PricedItem represents the total cost of that item

The Total now returns the sum of PricedItems. The PricedItems take each
BasketItem and calculate it's price.

The Price is calculated by either applying a discount, or if one is not
available, the standard price is returned.